requirements.txt
flask==2.0.1
werkzeug==2.0.1
python-dotenv==0.19.0
requests==2.31.0
google-api-python-client==2.108.0
beautifulsoup4==4.12.2